Fix NPE caused by call on uninitialized UICC slot
UiccController’s constructor is registering for multiple events,
including EVENT_RADIO_UNAVAILABLE, on both physical slots.
EVENT_RADIO_UNAVAILABLE arrives on slot 1 before mUiccSlots[1] is
initialized. The initialization happens in
UiccController#onGetIccCardStatusDone or
UiccController#onGetSlotStatusDone. In the single SIM case, this
initialization will only happen for mUiccSlots[0] and not for
mUiccSlots[1]. The exception causing code tries to retrieve the UiccSlot
for slot 1 without checking if it exists first.

Change-Id: Ie822f33c092eb96aa1cec2dc88bb23aa271c5b4e
CRs-Fixed: 3731563